refactor: extract useAuthSession from App.tsx (phase 4c)
Moves auth session state (session, authMode, credentials, auth flags),
two effects (subscribeToAuthChanges, setActiveSupabaseSession), and three
handlers (handleAuth, handlePasswordReset, handlePasswordUpdate) into
src/hooks/useAuthSession.ts.

The isPasswordResetMode and clearPasswordResetModeFromUrl URL utilities
move to the hook file where they belong. App.tsx gets clearPasswordResetModeFromUrl
back via hook return so handleSignOut can call it.

App.tsx: 717 → 568 lines. Behavior and all 658 tests unchanged.
